% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement0.R
\name{agreement0}
\alias{agreement0}
\title{Plot the agreement before recalibration}
\usage{
agreement0(object)
}
\arguments{
\item{object}{list returned by \link{measure_compare} function}
}
\description{
Plot the agreement and percentage agreement before recalibration along with
95\% confidence bands.
}
\examples{
### Load the data
data(data1)
### Analysis
measure_model <- measure_compare(data1)
### Plot the agreement without recalibration
agreement0(measure_model)
}
