% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreement1.R
\name{agreement1}
\alias{agreement1}
\title{Plot the agreement after recalibration}
\usage{
agreement1(object)
}
\arguments{
\item{object}{list returned by \link{measure_compare} function.}
}
\description{
Plot the agreement and percentage agreement after recalibration along with
95\% confidence bands.
}
\examples{
### Load the data
data(data1)
### Analysis
measure_model <- measure_compare(data1)
### Plot the agreement after recalibration
agreement0(measure_model)
}
