% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bias_plot.R
\name{bias_plot}
\alias{bias_plot}
\title{Plot the bias and measurements}
\usage{
bias_plot(object)
}
\arguments{
\item{object}{list returned by \link{measure_compare} function.}
}
\description{
Plot the bias and measurements from the two methods along with their
regression lines.
}
\examples{
### Load the data
data(data1)
### Analysis
measure_model <- measure_compare(data1)
### Plot the bias
bias_plot(measure_model)
}
