% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_compare.R
\name{measure_compare}
\alias{measure_compare}
\title{Process data and estimate models for method comparison}
\usage{
measure_compare(data, new = "y1", ref = "y2", id = "id", nb_simul = 1000)
}
\arguments{
\item{data}{a required data frame containing at least three columns: the
measurements for the two methods and ids.}

\item{new}{an optional string. The column name containing the measurements of the new
method.}

\item{ref}{an optional string. The column name containing the measurements of the
reference method (at least two measurements per individual).}

\item{id}{an optional string. The column name containing the ids of individuals.}

\item{nb_simul}{an optional number. The number of simulations used for confidence bands
simulations.}
}
\value{
A list containing
\itemize{
\item \code{models}: a list of models
\item \code{sub}: the original data frame containing with renamed columns and
additional computed data
\item \code{ref}: a data frame for the reference method and its related estimates
\item \code{y1_y2}: subset data frame of sub containing measurements for both
methods, that is \code{y1} and \code{y2} columns are not NAs
\item \code{new}: a data frame for the new method and its related estimates
\item \code{agg}: a data frame made from ref and new containing a row per
individual
\item \code{sim_params}: estimated model coefficients used afterward
\item \code{nb_simul}: the number of simulations used for confidence bands
simulations
\item \code{bias}: differential and proportional biases
}
}
\description{
\code{measure_compare()} implements the new estimation procedure to asses the
agreement between the two measurement methods.
}
\details{
Details of the function
}
\examples{
### Load the data
data(data1)
### Analysis
measure_model <- measure_compare(data1)
}
