% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mse.R
\name{mse}
\alias{mse}
\title{Plot the mean squared errors}
\usage{
mse(object)
}
\arguments{
\item{object}{list returned by \link{measure_compare} function}
}
\description{
Plot the mean squared errors from the two methods along with
95\% confidence bands.
}
\examples{
### Load the data
data(data1)
### Analysis
measure_model <- measure_compare(data1)
### Plot the bias
mse(measure_model)
}
