% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_agreement0.R
\name{pct_agreement0}
\alias{pct_agreement0}
\title{Plot the percentage agreement before recalibration}
\usage{
pct_agreement0(object)
}
\arguments{
\item{object}{list returned by \link{measure_compare} function}
}
\description{
Plot the percentage agreement before recalibration along with
95\% confidence bands.
}
\examples{
### Load the data
data(data1)
### Analysis
measure_model <- measure_compare(data1)
### Plot the percentage agreement without recalibration
pct_agreement0(measure_model)
}
