% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqrt_mse.R
\name{sqrt_mse}
\alias{sqrt_mse}
\title{Plot the square root of the mean squared errors}
\usage{
sqrt_mse(object)
}
\arguments{
\item{object}{list returned by \link{measure_compare} function.}
}
\description{
Plot the square root of the mean squared errors from the two methods along
with 95\% confidence bands.
}
\examples{
### Load the data
data(data1)
### Analysis
measure_model <- measure_compare(data1)
### Plot the bias
sqrt_mse(measure_model)
}
