% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_bias_plot.R
\name{total_bias_plot}
\alias{total_bias_plot}
\title{Plot total bias}
\usage{
total_bias_plot(object, object2 = NULL)
}
\arguments{
\item{object}{list returned by \link{measure_compare} function.}

\item{object2}{(optional) returned by \link{measure_compare} function.
If provided, will plot a second total bias estimate.}
}
\description{
Plot total bias along with 95\% confidence bands
}
\examples{
### Load the data
data(data1)
### Analysis
measure_model <- measure_compare(data1)
### Plot the total bias
total_bias_plot(measure_model)
}
