% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_plot.R
\name{compare_plot}
\alias{compare_plot}
\title{Plot used to visualize the recalibration of the new method after estimating
the bias}
\usage{
compare_plot(object)
}
\arguments{
\item{object}{list returned by \link{measure_compare} function.}
}
\description{
This function allows the visualization of the bias-corrected values (i.e.
recalibrated values, variable y1_corr) of the new measurement method.
}
\examples{
\donttest{
### Load the data
data(data1)
### Analysis
measure_model <- measure_compare(data1, nb_simul=100)
### Plot the bias
compare_plot(measure_model)}
}
