% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_plot.R
\name{precision_plot}
\alias{precision_plot}
\title{Plot the precision of the methods}
\usage{
precision_plot(object)
}
\arguments{
\item{object}{list returned by \link{measure_compare} function.}
}
\description{
Plot the precision of the two methods along with their 95\% confidence bands
}
\examples{
### Load the data
data(data1)
### Analysis
measure_model <- measure_compare(data1)
### Plot the precision of the two methods
precision_plot(measure_model)
}
